chore: experiment with json-rpc-2.0 lib
Seems like this could be a reasonable way to manage message passing
between extension components. Using chrome.runtime.Port for messaging
would be desirable as it will notify when the endpoint is disconnected.
But it doesn't provide request-response semantics, just message passing.
json-rpc-2.0 can add that, plus a bit of structure. And JSON RPC is
commonly used for Ethereum things.