feat: set the reddit-side component
When we run on a Reddit tab, we inject a content script into the tab
to handle communication with Reddit APIs on behalf of the logged-in
user.

I may change the tab-sharing model, but in the future, but currently we
allow 1 reddit tab to host communication with more than one of our UIs.
Closing the tab will cut off the UI's connectivity, so we show a modal
while we're active to make it clear that the tab is being used.

Probably need to revisit this setup, as it has the potential to be
confusing for users if they don't have a mental model of this
connectivity between extension UI and tab, especially when the in-use
reddit tab may not be visible (when a second UI opens some time after
the first).