build: use swc to speed up rollup build
We use @rollup/plugin-swc instead of @rollup/plugin-typescript to
transform our source files. The builds were starting to get kinda slow,
and this seems to help. The downside is that we don't type-check ts at
build time, but we can still run tsc to do that, either manually or via
run commands in package.json.