build: use lib mode for predictable entry point interfaces
We were building with a mix of HTML and JS entry points. The reddit
entry point was set up in an unpredictable way by Vite, which made in
not execute correctly using the dynamic import() from
browser.scripting.executeScript(). We now build all JS in lib mode,
which results in the entrypoints exporting our exports, rather than
running for side effect.