refactor!: change purpose of getAccountVaultAddress
getAccountVaultAddress was used to get the address of the active
session's account. It's now used to get the vault address of the account
identified by a specific username.

This is changing because there is a different API that provides more
details about a user's own account's vault addresses (past and present),
we'll use that for user details.

We don't currently have a UI design to expose a feature to look up the
vault address of another Reddit user, so this call will not be used
anywhere for now.