feat: use JSONRPC between UI and Background
We're now using JSONRPC to communicate between the UI and the Background
service. The UI no longer connects directly to Reddit tabs, instead the
Background service creates Reddit tab connections, and proxies them
through the UI <-> Background connection.

This approach provides more structure than the previous approach, which
used extension messaging to communicate the availability of a Reddit
tab. The dev server can't connect directly to a Reddit tab, and the dev
server can't send or receive extension messages without additional
proxying via a content script. We allow dev-server pages to connect to
the dev-mode extension with an external Port connection, using a
externally_connectable rule in manifest.json.

This provides a consistent communication method for both the regular
sidebar extension UI, and dev-server pages. It also provides a clear
communication framework which can be easily extended in the future to
support additional messages.