feat: implement reddit/private-cache
This allows us to persistently store data private to the user's reddit
account in such a way that it's only accessible to our content script
(running in the reddit web page), not reddit's web page itself, and not
the rest of the extension.

We'll initially use this to cache the user's API credentials to avoid
needing to re-fetch them regularly.