feat!: add userId assertion param to reddit RPC calls
Our Reddit RPC calls now take a userId param that allows the client to
assert that the request is intended for the account with that userId.

This protects against a race condition where the UI makes a
state-changing request intended for one account, that affects a
different account, because the session changed on the Reddit side.