chore: add RawEthAddress zod type
It validates a 40 hex digit eth address, without changing it to to a
checksum address.

I created this for use with EIP-712 data, but now that I think about it,
the message hash shouldn't be affected by an address value being in
checksum format, so perhaps we don't need to maintain the address case.