feat: rework session caching to facilitate revalidation
The reddit-side session caching behaviour is now able to selectively
bypass the persistent cache in order to revalidate the
currently-logged-in user.

When a request for a session includes the expected user ID, we allow the
response to be served from the persistent cache or in-memory cache. When
a request does not specify a user ID, we re-validate by making a request
to reddit for the currently-logged-in user's session.

This allows the UI to control session freshness. The UI can make two
requests on startup — one for the last-active user, which should be
served quickly from the cache. And another for the currently-logged-in
user. In this way, it can quickly serve the user's session in most
cases, but also notice that the user has changed, without additional
effort.