fix: stop binding Port multiple times
The Reddit RPC server was being bound to the same port multiple times
when the extension was activate more than once on a single tab. This was
because handleRedditInteractionConnections() was registering a listener
to browser.runtime.onConnect that was not removed when stop() was
called, so the next start() would result in two listeners being bound;
and then more on subsequent stop/start cycles.

This resulted in multiple messages being handled by the JSON RPC server
— it seems de-duplicate responses, but not requests.

browser.runtime.onConnect.addListener