refactor: fix UI connecting to background twice
We were using a useRef to create and hold the connection to the
background service, but weren't accounting for useRef creating the ref
twice in dev mode. We now create the connection in a useEffect hook with
correct teardown, and store the provider in the UI state. Same for the
redditProvider.