feat: adjust UI for minimum screen size
We now set a minimum width of 385px, below this the UI gets
proportionally scaled; above is responsive. The sidebar seems to have a
min-width of 320px. We could adjust the UI with a responsive layout to
fit into 320 directly, but I don't think it's worth the effort &
maintenance burdon, considering 320 is so close to 385 which works well.