feat: scale UI down at smallest breakpoint
We now scale the UI down at the smallest breakpoint to fit it into a
320px sidebar (the minimum width). For some reason, setting a min width
using the HTML's meta viewport does not work in the extension sidebar.
That would be the ideal way to progressively scale down the UI when the
available width is less than the reasonable minimum.

Because we can't do that, instead we scale the rem units down at the
smallest breakpoint to shrink the UI. This has much the same effect,
except for a slight jump at the breakpoint, rather than smooth scaling.