fix: remove all trace of help footer when it's closed
We were keeping an invisible element in the page after the help footer
was closed, which resulted in a long empty space when closing the
expandable pairing area.

I've factored out the reserved space component from the rest of the help
footer, an adjusted it to be display:none after it's faded out.