fix: work around wagmi impure-render bug / disconnect
Toggling our inline help on/off was resulting in Wagmi disconnecting
its account, and logging a warning about it changing state while
rendering.

The warning seems to be a Wagmi bug. Not sure if Wagmi expects to never
re-render after mounting. In any case, we can avoid this by moving the
help state to be a descendant of the WagmiProvider, to stop help
triggering Wagmi re-renders.