wip: feat: track reddit logged-out in UI state
We now maintain a flag in the UI state which is true while Reddit seems
to be logged out. We watch RedditProvider requests, and set the flag
when a request fails due to being logged out, and un-set the flag when
a request succeeds or fails for any other reason.