refactor: make SearchForUser a query rather than mutation
I think this is more natural. The query parsing is now done at the UI
level rather than when the search executes. The UI needs to validate the
search, so it is more natural to let the UI use the parser to validate
the user input, and this removes a failure mode of the search process.