feat: support lazy-reconnecting Port connections
Previously we've relied on Ports staying connected, but in practice the
background service worker gets stopped after ~30s unless it's active.
It'd be possible to send keep-alive messages, but probably best to not
fight the system, and design things in such a way as to work well with
temporary connections that are created when required and stopped when
not.

To support this, we now use a pair of abstractions: Connector<T> and
ManagedConnection<T>. Connector<T> is the simpler of the two: a function
that returns a pair of [connection: T, disconnectConnection: () => void].
So it generalises creating a connection and being able to disconnect it.
The Connector function also accepts a callback that will be notified if
the returned connection disconnects.

ManagedConnection is a stateful object that uses a Connector to create
connections on demand.