refactor: use on-demand connections for bg services
The BackgroundService and VaultonomyBackgroundServiceSession are now
implemented using the ManagedConnection / Connector abstraction, which
means they can now deal with Reddit tabs going away and being replaced
more gracefully.

We no longer maintain an "availability" connection from the Reddit tab
to the backend, because the main reason I did that was to use it to
monitor explicit disconnects from the Reddit tab, and (initially) as a
keep-alive mechanism. But because we're going to not show a UI in Reddit
tabs now, and we don't want a keep-alive as we expect/want the bg to
shutdown when not  in use, it's unnecessary.