feat: implement new Reddit tab availability mechanism
We now determine whether a Reddit tab is available to connect to using
webextension tabs APIs to query and monitor tab events. This gives us
a more accurate and timely view of connectivity, as we can track
multiple candidates to connect with.

Previously we expected to have a single tab locked and dedicated to
extension connectivity, so we used a Port connection from that one tab
to determine connectivity. That model doesn't mesh with the current
approach of not locking tabs that are used for extension requests.