chore: resolve TODO on scripting.executeScript
The method we're using to build the content script as an ESM library
with an explicit default export, and import it from a regular JS file
is fine. Specifying a function here seems possible, but it's fiddly and
doesn't really have any benefit over the tiny static ESM loader we have
in reddit-contentscript.js.