feat: improve user search responsiveness
Searching for users is now faster, both when manually and automatically
searching from Reddit interactions. We achieve this by prefetching
queries.

For automatic queries, we prefetch the user when we first start seeing
interest in them before it's confirmed a few hundred ms later. This
results in their details being available by the time they've been
hovered long enough to confirm.

For manual queries, we now prefetch the user vault when searching for a
username, so the vault details are typically available when the user
profile loads.