feat: polish user auto-search
We now clear the auto-search when clearing the search input. We show a
placeholder value in the search input for the user being displayed.
(Could do something more in the UI here to call out the auto-search
behaviour, but maybe this is enough.)

And the code refers the the auto search as userOfInterest, which feels
more descriptive.