feat: wake up UI when bg starts from a user event
When the background service worker is stopped and a user triggers a user
interaction event, the background wakes up and tries to send the event
to all connected UIs via RPC connections. But when the bg has been
stopped, all the RPC connections are stopped, so the UIs wouldn't
receive the event notifications.

Using extension messages to broadcast notifications would avoid this,
but it wouldn't work for the dev-mode UI, as that can't receive
extension messages (which is why we have the RPC notifications in the
first place).

So to communicate the events to UIs, the bg service fires a single
extension message to notify the UI that it's started, and the UIs then
make a regular RPC request to request a list of events that have been
queued since the bg started. The UI can then emit those events and carry
on as normal.