feat: stop the reddit interaction content script
The Reddit interaction server that runs as a content script in a Reddit
tab now auto-stops after a short period of inactivity.

This is necessary to clean up after stopping/uninstalling the extension
(Chrome seems to not actively stop content scripts when disabling
extensions), and also acts as a safeguard against wastefully running too
many instances (although we should only run one at once, as the
background service attempts to reuse the same tab).