feat: detect whether the user has disabled global reddit access
We request global access to reddit tabs by default, but users can choose
to remove this access in extension options. In that case we still work
via activeTab access to the current reddit tab (if any), but we need to
detect this condition to customise error messages when no reddit tab is
available.