feat: implement cross-origin fetch
We need to be able to make HTTP requests to Reddit API endpoints that
are not CORS enabled. Unfortunately this isn't possible from the content
script sandboxed in the Reddit tab — we can only make non-CORS requests
from the background service worker.

To make these requests while trying to avoid giving the main extension
access to the user's API auth token, we now have a separate extension
message that just performs HTTP requests to the Reddit API endpoint
without CORS restrictions. Auth token credentials are passed in the
request headers, but the background doesn't access them or log
credentials.

Maybe It'd make sense to drop the web page "sandbox" and explicitly pass
the user auth token from the page to the background worker and do all
the requests from there... 🤷