feat: show rejected pairing request error
When Reddit rejects a pairing message (via an API error) we now show an
error message indicating this with a possible resolution, rather than a
more generic error message.

This error is likely to occur when users attempt to pair an address with
multiple accounts, but there could be other causes.