test: test useImageSize hook
UserAvatar used to include logic to detect the size of a loaded image,
but we previously refactored it out into a separate hook. The UserAvatar
test was only testing the image loading logic, and the rest of the
UserAvatar logic relies on visual appearance to verify, so a unit test
does not seem particularly useful.