build: remove bundled code that calls eval()
Vite was warning that code in the bundle calls eval(). eval() is not
permitted in Chrome Web Store extensions. The code calling eval() was
from a polyfill of the vm module. We don't use it, but was
unintentionally being included by nodePolyfills() vite plugin. (Because
it interprets an empty list of modules to include as "include all
modules".) We now only polyfill Buffer and process globals, which are
used by various MetaMask-related dependencies.