build(ci): automate submission of releases to AMO
scripts/amo-submit.sh can synchronise a GitHub Release (created by the
ci.yml workflow) to Firefox Add-ons (AMO). We download the signed .xpi
file, attest to it and attach it to the GitHub Release. This will allow
people to verify the whole chain — from build (in ci.yml) to submission
& signing in deploy_amo.yml — using the attestations rather than needing
to extract and verify checksums in the archives.

Mozilla do provide a tool to upload to AMO[1], but it doesn't fit well
with our requirements, as it seems to expect to package the extension
itself, it doesn't have an option to upload an existing archive. We also
want to validate the provenance attestation of our release archive
before uploading, so we'd have to script that anyway. So all in all it
seems fine to do a few API calls ourself for greater control.

[1]: https://extensionworkshop.com/documentation/develop/web-ext-command-reference/#web-ext-sign

The listed releases will time out without fetching the .xpi initially,
as we need to be manually verified by AMO before we can list on AMO.