wip: build: automate submission of releases to AMO
Support release and pre-release submissions to AMO. For pre-releases we
wait for and download the signed .xpi file and attach it to the GitHub
Release.

For normal releases we only submit and don't wait for approval,
as approval can require manual intervention (at least for new
extensions). We could also re-run this later and fetch the signed stable
.xpi, but not sure that it's necessary.