build(ci): submit releases to Chrome Web Store
We now automate the submission of releases to the Chrome Web Store. We
have two items in CWS, the default stable version, and a pre-release
unstable version. The pre-release version has both stable and
pre-release versions deployed, whereas the default version only gets
stable releases.

The pre-release item will be unlisted and is intended for testing.