feat: show "Not connected" on Vault heading
We now show a "Not connected" subtitle on the Reddit Vault heading when
Reddit is not connected, like we do when the Wallet is not connected.

We are still showing the cached Vault data for the user, which may be
confusing, but I'm leaning towards keeping it, as it hints that
disconnection is a temporary/normal situation that can fixed by just
opening a Reddit tab. Probably needs some user testing.